chore(repo): stage and sync multiple compatibility + engine updates
- Stage and commit modified compatibility headers in Core and GeneralsMD
- Include several engine/source updates (W3DDevice, texture loader, CMakeLists)
- Add a few missing compatibility headers as tracked files

Refs: Phase 50 compatibility alignment

+++ GeneralsMD/Code/GameEngineDevice/Source/W3DDevice/GameClient/W3DAssetManager.cpp
@@ -771,7 +771,7 @@ RenderObjClass * W3DAssetManager::Create_Render_Obj(
 		const char *mesh_name = strchr (name, '.');
 		if (mesh_name != NULL)
 		{
-			lstrcpyn(filename, name, ((int)mesh_name) - ((int)name) + 1);
+			lstrcpyn(filename, name, (int)(mesh_name - name) + 1);
 			lstrcat(filename, ".w3d");
 		} else {
 			sprintf( filename, "%s.w3d", name);
@@ -1371,7 +1371,7 @@ RenderObjClass * W3DAssetManager::Create_Render_Obj(const char * name,float scal
 		char filename [MAX_PATH];
 		char *mesh_name = ::strchr (name, '.');
 		if (mesh_name != NULL) {
-			::lstrcpyn (filename, name, ((int)mesh_name) - ((int)name) + 1);
+			::lstrcpyn (filename, name, (int)(mesh_name - name) + 1);
 			if (isGranny)
 				::lstrcat (filename, ".gr2");
 			else
